Moves stores api to service layer
